+++ src/iouring_op_batch_rep.cc
@@ -35,7 +35,7 @@ void op_batch_rep::set_user_data(void* userdata)
         ::io_uring_sqe_set_data(&sqe, userdata);
 }
 
-void op_batch_rep::set_user_data(uint64_t userdata)
+void op_batch_rep::set_user_data(uintptr_t userdata)
 {
     for (auto& sqe : m_sqes)
         ::io_uring_sqe_set_data64(&sqe, userdata);
